requirements.txt
Flask
psycopg2-binary
gunicorn
python-dotenv
Flask-CORS
Flask-JWT-Extended